Bloom/requirements.txt
streamlit>=1.0
numpy<2.0
pandas>=1.5.0
pyarrow>=10.0.1
numexpr>=2.8.0
bottleneck>=1.3.4
jinja2>=3.1.2
pybind11>=2.12
python-binance>=1.0.15
importnb==2023.11.1
